Fix form input collection bug.
Since each form input is wrapped in a box that contains the input as
well as the validation error message, passing `collect=False` didn't
behave as expected. This fix forwards the `collect` kwarg to the
wrapper box and overrides `collect()` on the form input to call
`collect()` on the box.

Fixes https://github.com/hyperdiv/hyperdiv/issues/52