Event props cannot be read in tasks
Event props are set for one run of the main app function and reset at
the end of that run. Tasks which run asynchronously will not be likely
to ever read the non-default value of an event prop when it is set, so
attempting to read an event prop will likely not behave as expected.